+++ src/sections/Abstract.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
-import SectionBorder from './SectionBorder.vue'
-import SectionTitle from './SectionTitle.vue'
-import Freetext from './Freetext.vue'
+import SectionBorder from '../components/SectionBorder.vue'
+import SectionTitle from '../components/SectionTitle.vue'
+import Freetext from '../components/Freetext.vue'
 </script>
 
 <template>