+++ src/sections/IdentifierEsri.vue
@@ -3,8 +3,8 @@ import { computed, type ComputedRef, ref, watch } from 'vue'
 
 import type { Identifier } from '../types/iso'
 
-import FormLabel from './FormLabel.vue'
-import FormInput from './FormInput.vue'
+import FormLabel from '../components/FormLabel.vue'
+import FormInput from '../components/FormInput.vue'
 
 const account = 'bas'