+++ src/sections/Resources.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
-import SectionBorder from './SectionBorder.vue'
-import SectionTitle from './SectionTitle.vue'
-import Link from './Link.vue'
+import SectionBorder from '../components/SectionBorder.vue'
+import SectionTitle from '../components/SectionTitle.vue'
+import Link from '../components/Link.vue'
 </script>
 
 <template>