+++ src/sections/Services.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
 import servicesData from '../data/services.json'
 
-import SectionBorder from './SectionBorder.vue'
-import SectionTitle from './SectionTitle.vue'
+import SectionBorder from '../components/SectionBorder.vue'
+import SectionTitle from '../components/SectionTitle.vue'
 import Service from './Service.vue'
 
 const serviceSlugs: string[] = Object.keys(servicesData.services).sort((a: string, b: string) =>